dom/tests/mochitest/general/test_497898.html

<html>
<head>
<title>Crash [@ nsFocusManager::SendFocusOrBlurEvent] after switching focus to a different window in this case</title>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script>
SimpleTest.waitForExplicitFinish();

function done()
{
  is("passed", "passed", "test passed without crashing");
  SimpleTest.finish();
}

function switchFocus()
{
  setTimeout(() => window.open('497633.html', '_new', 'width=300,height=300'), 0);
}
</script>
</head>
<body>
  <iframe srcdoc="<html><meta charset='utf-8'>
      <head>
        <script>
          function run() {
            setTimeout(() => document.getElementById('a').focus(), 0);
          }
        </script>
      </head>
      <body onload='run()'>
        <button id='a' onfocus='parent.switchFocus()' onblur='window.frameElement.parentNode.removeChild(window.frameElement)'>Switching focus to a different program should not crash Mozilla</button>
      </body>
    </html>">
  </iframe>

<p id="display"></p>
<div id="content" style="display: none"></div>

</body>
</html>
